Strict user validation, display category in events list

+++ src/main/java/info/jawne/kalendarz/controllers/CategoryDeleteController.java
@@ -2,6 +2,8 @@ package info.jawne.kalendarz.controllers;
 
 import java.util.Locale;
 
+import javax.servlet.http.HttpSession;
+
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,7 +16,10 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import info.jawne.kalendarz.dao.CategoryDao;
+import info.jawne.kalendarz.dao.UserDao;
+import info.jawne.kalendarz.exceptions.AuthorizationException;
 import info.jawne.kalendarz.models.Message;
+import info.jawne.kalendarz.models.User;
 
 @Controller
 public class CategoryDeleteController {
@@ -26,8 +31,16 @@ public class CategoryDeleteController {
 	@Autowired
 	private ReloadableResourceBundleMessageSource messageSource;
 
+	@Autowired
+	private UserDao user_dao;
+
 	@RequestMapping(value = "/categories/{id}/delete", method = RequestMethod.POST)
-	public String action(Model model, @PathVariable int id, RedirectAttributes redirectAttributes) {
+	public String action(Model model, @PathVariable int id, RedirectAttributes redirectAttributes, HttpSession session)
+			throws AuthorizationException {
+		User user = user_dao.getByUsernameOrNull((String) session.getAttribute("username"));
+		if (user == null) {
+			throw new AuthorizationException();
+		}
 		category_dao.remove(category_dao.getById(id));
 		redirectAttributes.addFlashAttribute("message", new Message(Message.Status.SUCCESS,
 				messageSource.getMessage("CategoryFormController.saved", null, Locale.US)));
@@ -36,7 +49,11 @@ public class CategoryDeleteController {
 	}
 
 	@RequestMapping(value = "/categories/{id}/delete", method = RequestMethod.GET)
-	public String confirmation(Model model, @PathVariable long id) {
+	public String confirmation(Model model, @PathVariable long id, HttpSession session) throws AuthorizationException {
+		User user = user_dao.getByUsernameOrNull((String) session.getAttribute("username"));
+		if (user == null) {
+			throw new AuthorizationException();
+		}
 		model.addAttribute("category", category_dao.getById(id));
 		return "categoryDelete";
 	}
